+++ cmd_queue/__init__.py
@@ -296,7 +296,7 @@ Example:
 __mkinit__ = """
 mkinit -m cmd_queue
 """
-__version__ = '0.1.9'
+__version__ = '0.1.10'
 
 
 __submodules__ = {