+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.1.10 - Unreleased
 
+### Added
+* the `cli_boilerplate` submodule for help writing consistent scriptconfig + cmdqueue CLIs
+* util yaml
+
 
 ## Version 0.1.9 - Released 2023-04-04