+++ CHANGELOG.md
@@ -5,7 +5,14 @@ We are currently working on porting this changelog to the specifications in
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Version 0.1.9 - Unreleased
+## Version 0.1.10 - Unreleased
+
+### Added
+* the `cli_boilerplate` submodule for help writing consistent scriptconfig + cmdqueue CLIs
+* util yaml
+
+
+## Version 0.1.9 - Released 2023-04-04
 
 ### Added
 * Support for more sbatch options in slurm backend